+++ controller/src/main/java/com/linbit/linstor/core/LinstorDatabaseTool.java
@@ -28,11 +28,13 @@ import com.linbit.linstor.debug.DebugModule;
 import com.linbit.linstor.event.EventModule;
 import com.linbit.linstor.event.handler.EventHandler;
 import com.linbit.linstor.event.handler.protobuf.controller.ConnectionStateEventHandler;
+import com.linbit.linstor.event.handler.protobuf.controller.DonePercentageEventHandler;
 import com.linbit.linstor.event.handler.protobuf.controller.ReplicationStateEventHandler;
 import com.linbit.linstor.event.handler.protobuf.controller.ResourceStateEventHandler;
 import com.linbit.linstor.event.handler.protobuf.controller.VolumeDiskStateEventHandler;
 import com.linbit.linstor.event.serializer.EventSerializer;
 import com.linbit.linstor.event.serializer.protobuf.common.ConnectionStateEventSerializer;
+import com.linbit.linstor.event.serializer.protobuf.common.DonePercentageEventSerializer;
 import com.linbit.linstor.event.serializer.protobuf.common.ReplicationStateEventSerializer;
 import com.linbit.linstor.event.serializer.protobuf.common.ResourceStateEventSerializer;
 import com.linbit.linstor.event.serializer.protobuf.common.VolumeDiskStateEventSerializer;
@@ -182,12 +184,14 @@ public class LinstorDatabaseTool
             ResourceStateEventSerializer.class,
             VolumeDiskStateEventSerializer.class,
             ReplicationStateEventSerializer.class,
+            DonePercentageEventSerializer.class,
             ConnectionStateEventSerializer.class
         );
         List<Class<? extends EventHandler>> eventHandlers = Arrays.asList(
             ResourceStateEventHandler.class,
             VolumeDiskStateEventHandler.class,
             ReplicationStateEventHandler.class,
+            DonePercentageEventHandler.class,
             ConnectionStateEventHandler.class
         );
         final List<Module> injModList = new LinkedList<>(