+++ server/src/main/java/com/linbit/linstor/api/interfaces/serializer/CommonSerializer.java
@@ -80,6 +80,7 @@ public interface CommonSerializer
 
         CommonSerializerBuilder volumeDiskState(String diskState);
         CommonSerializerBuilder replicationState(String replicationState);
+        CommonSerializerBuilder donePercentageEvent(Float donePercentage);
 
         CommonSerializerBuilder resourceStateEvent(ResourceState resourceState);