+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Notify controller about replication state changes in DRBD
+- Notify controller about done percentage changes in DRBD
 
 ## [1.30.4] - 2025-02-03